setting up homepage
